Working on first function
After playing around I found out why the hint of template literals was given, I can get the element to display with no information fed in (makes sense, no data object is being pulled) it doesn't work inside the fuction. I'll need to work on that and figure out why, I think it has to do with the second if statement to check the index of the items.